Renamed the Elog package to sam
