+++ internal/sam/texter.go
@@ -1,4 +1,4 @@
-package elog
+package sam
 
 import (
 	"github.com/rjkroege/edwood/internal/util"
