+++ regx_test.go
@@ -2,9 +2,10 @@ package main
 
 import (
 	"fmt"
-	"github.com/rjkroege/edwood/internal/elog"
 	"reflect"
 	"testing"
+
+	"github.com/rjkroege/edwood/internal/sam"
 )
 
 func TestRegexpForward(t *testing.T) {
@@ -25,7 +26,7 @@ func TestRegexpForward(t *testing.T) {
 			if err != nil {
 				t.Fatalf("failed to compile regular expression %q", tc.re)
 			}
-			text := elog.NewTextBuffer(0, 0, []rune(tc.text))
+			text := sam.NewTextBuffer(0, 0, []rune(tc.text))
 			rs := re.rxexecute(text, nil, 0, text.Nc(), tc.nmax)
 			if !reflect.DeepEqual(rs, tc.expected) {
 				t.Errorf("regexp %q incorrectly matches %q:\nexpected: %v\ngot: %v",
@@ -54,7 +55,7 @@ func TestRegexpBackward(t *testing.T) {
 			if err != nil {
 				t.Fatalf("failed to compile regular expression %q", tc.re)
 			}
-			text := elog.NewTextBuffer(0, 0, []rune(tc.text))
+			text := sam.NewTextBuffer(0, 0, []rune(tc.text))
 			rs := re.rxbexecute(text, text.Nc(), tc.nmax)
 			if !reflect.DeepEqual(rs, tc.expected) {
 				t.Errorf("regexp %q incorrectly matches %q:\nexpected: %v\ngot: %v",
